Removing unnecessary typeclasses and refactoring StreamLoop.
