+++ core/src/main/scala/com/github/gvolpe/fs2rabbit/StreamLoop.scala
@@ -16,9 +16,8 @@
 
 package com.github.gvolpe.fs2rabbit
 
-import cats.effect.Effect
-import com.github.gvolpe.fs2rabbit.typeclasses.{EffectScheduler, EffectUnsafeSyncRunner}
-import fs2.Stream
+import cats.effect.{Effect, IO}
+import fs2.{Scheduler, Stream}
 import org.slf4j.LoggerFactory
 
 import scala.concurrent.ExecutionContext
@@ -36,17 +35,21 @@ object StreamLoop {
 
   private val log = LoggerFactory.getLogger(getClass)
 
-  def run[F[_] : Effect : EffectScheduler : EffectUnsafeSyncRunner](program: () => Stream[F, Unit], retry: FiniteDuration = 5.seconds)
-         (implicit ec: ExecutionContext): Unit = {
-    EffectUnsafeSyncRunner[F].unsafeRunSync(loop(program(), retry).run)
-  }
+  def run[F[_]](program: () => Stream[F, Unit],
+                retry: FiniteDuration = 5.seconds)
+               (implicit F: Effect[F], ec: ExecutionContext): IO[Unit] =
+    F.runAsync(loop(program(), retry).run) {
+      case Right(_) => IO.unit
+      case Left(e)  => IO.raiseError(e)
+    }
 
-  private def loop[F[_] : Effect : EffectScheduler](program: Stream[F, Unit], retry: FiniteDuration)
-                        (implicit ec: ExecutionContext): Stream[F, Unit] = {
+  private def loop[F[_] : Effect](program: Stream[F, Unit], retry: FiniteDuration)
+                                 (implicit ec: ExecutionContext): Stream[F, Unit] = {
     program.handleErrorWith { err =>
       log.error(s"$err")
       log.info(s"Restarting in $retry...")
-      loop[F](Stream.eval(EffectScheduler[F].schedule[Unit](program.run, retry)), retry)
+      val scheduledProgram = Scheduler[F](2).flatMap(_.sleep[F](retry)).flatMap(_ => program)
+      loop[F](scheduledProgram, retry)
     }
   }
