+++ core/src/main/scala/com/github/gvolpe/fs2rabbit/utils/IOApp.scala
@@ -14,19 +14,11 @@
  * limitations under the License.
  */
 
-package com.github.gvolpe.fs2rabbit.examples.scheduler
+package com.github.gvolpe.fs2rabbit.utils
 
-import com.github.gvolpe.fs2rabbit.typeclasses.EffectScheduler
-import monix.eval.Task
+import cats.effect.IO
 
-import scala.concurrent.ExecutionContext
-import scala.concurrent.duration.FiniteDuration
-
-object MonixEffectScheduler extends EffectScheduler[Task] {
-
-  override def schedule[A](effect: Task[A], delay: FiniteDuration)
-                          (implicit ec: ExecutionContext): Task[A] = {
-    effect.delayExecution(delay)
-  }
-
+trait IOApp {
+  def start(args: List[String]): IO[Unit]
+  def main(args: Array[String]): Unit = start(args.toList).unsafeRunSync()
 }
