+++ examples/src/main/scala/com/github/gvolpe/fs2rabbit/examples/IODemo.scala
@@ -19,16 +19,17 @@ package com.github.gvolpe.fs2rabbit.examples
 import cats.effect.IO
 import com.github.gvolpe.fs2rabbit.StreamLoop
 import com.github.gvolpe.fs2rabbit.interpreter.Fs2RabbitInterpreter
+import com.github.gvolpe.fs2rabbit.utils.IOApp
 
 import scala.concurrent.ExecutionContext
 
-object IODemo extends App {
+object IODemo extends IOApp {
 
   implicit val appS: ExecutionContext = scala.concurrent.ExecutionContext.Implicits.global
   implicit val interpreter: Fs2RabbitInterpreter[IO] = new Fs2RabbitInterpreter[IO]
 
-  val program = new GenericDemo[IO]().program
-
-  StreamLoop.run(() => program)
+  override def start(args: List[String]): IO[Unit] = {
+    StreamLoop.run(() => new GenericDemo[IO]().program)
+  }
 
 }
