+++ examples/src/main/scala/com/github/gvolpe/fs2rabbit/examples/MonixTaskDemo.scala
@@ -16,20 +16,22 @@
 
 package com.github.gvolpe.fs2rabbit.examples
 
+import cats.effect.IO
 import com.github.gvolpe.fs2rabbit.StreamLoop
 import com.github.gvolpe.fs2rabbit.interpreter.Fs2RabbitInterpreter
+import com.github.gvolpe.fs2rabbit.utils.IOApp
 import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
 
 import scala.concurrent.ExecutionContext
 
-object MonixTaskDemo extends App {
+object MonixTaskDemo extends IOApp {
 
   implicit val appS: ExecutionContext = scala.concurrent.ExecutionContext.Implicits.global
   implicit val interpreter: Fs2RabbitInterpreter[Task] = new Fs2RabbitInterpreter[Task]
 
-  val program = new GenericDemo[Task].program
-
-  StreamLoop.run(() => program)
+  override def start(args: List[String]): IO[Unit] = {
+    StreamLoop.run(() => new GenericDemo[Task].program)
+  }
 
 }
