--- examples/src/main/scala/com/github/gvolpe/fs2rabbit/examples/package.scala
@@ -1,29 +1,0 @@
-/*
- * Copyright 2017 Fs2 Rabbit
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.gvolpe.fs2rabbit
-
-import com.github.gvolpe.fs2rabbit.examples.runner.{IOEffectRunner, MonixEffectRunner}
-import com.github.gvolpe.fs2rabbit.examples.scheduler.{IOEffectScheduler, MonixEffectScheduler}
-
-package object examples {
-
-  implicit val iOEffectScheduler    = IOEffectScheduler
-  implicit val ioEffectRunner       = IOEffectRunner
-  implicit val monixEffectScheduler = MonixEffectScheduler
-  implicit val monixEffectRunner    = MonixEffectRunner
-
-}
