--- examples/src/main/scala/com/github/gvolpe/fs2rabbit/examples/scheduler/IOEffectScheduler.scala
@@ -1,33 +1,0 @@
-/*
- * Copyright 2017 Fs2 Rabbit
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.gvolpe.fs2rabbit.examples.scheduler
-
-import cats.effect.IO
-import com.github.gvolpe.fs2rabbit.typeclasses.EffectScheduler
-import fs2.Scheduler
-
-import scala.concurrent.ExecutionContext
-import scala.concurrent.duration.FiniteDuration
-
-object IOEffectScheduler extends EffectScheduler[IO] {
-
-  override def schedule[A](effect: IO[A], delay: FiniteDuration)
-                          (implicit ec: ExecutionContext): IO[A] = {
-    Scheduler[IO](2).flatMap(_.sleep[IO](delay)).run.flatMap(_ => effect)
-  }
-
-}
